refs CRASM-3542: Restructure RiskyOpenServices into modular component architecture
- Break out RiskyOpenServices into separate component files
- Create dedicated files for types, styles, and sub-components
- Organize imports into logical sections (React, MUI, hooks, components, types)
- Add JSDoc documentation headers
