+++ frontend/src/pages/VulnerabilityScanDash/RiskyOpenServices.tsx
@@ -1,194 +1,2 @@
-import React from 'react';
-import Box from '@mui/material/Box';
-import Divider from '@mui/material/Divider';
-import Grid from '@mui/material/Grid';
-import Table from '@mui/material/Table';
-import TableBody from '@mui/material/TableBody';
-import TableCell from '@mui/material/TableCell';
-import TableHead from '@mui/material/TableHead';
-import TableRow from '@mui/material/TableRow';
-import { ServiceData } from 'types/vuln-scan-stats';
-import { useAuthContext } from 'context';
-import CustomAlert from '@/components/Dashboard/CustomAlert';
-import { formatDisplayValue } from '@/utils/stringUtils';
-
-interface RiskyOpenServicesProps {
-  data: ServiceData[];
-  detectedServices?: string | number;
-  isKeyMetricsNull?: boolean;
-}
-
-const tableRowStyles = {
-  borderRadius: 5,
-  borderColor: 'gray',
-  '& td': {
-    borderLeft: '1px solid #ccc',
-    backgroundColor: '#fff',
-    py: '5px',
-    px: 2
-  },
-  '& td:first-of-type': {
-    borderTopLeftRadius: 8,
-    borderBottomLeftRadius: 8
-  },
-  '& td:last-of-type': {
-    borderTopRightRadius: 8,
-    borderBottomRightRadius: 8
-  }
-};
-
-function RiskyServicesTable({
-  data,
-  placement = 'top'
-}: {
-  data: ServiceData[];
-  placement?: string;
-}) {
-  const tableHeaderStyles = {
-    '& th': {
-      border: 'none',
-      backgroundColor: 'transparent',
-      pb: 0,
-      fontSize: '12px'
-    }
-  };
-  return (
-    <Table
-      sx={{
-        borderCollapse: 'separate',
-        borderSpacing: '0 10px',
-        'thead tr:first-of-type th': {
-          pt: 0
-        },
-        tableLayout: 'fixed',
-        width: '100%'
-      }}
-      component="table"
-    >
-      <TableHead component="thead">
-        {placement === 'top' ? (
-          <TableRow sx={tableHeaderStyles} component="tr">
-            <TableCell sx={{ minWidth: '147px' }} component="th" scope="col">
-              Service Name
-            </TableCell>
-            <TableCell
-              sx={{ width: { md: 'auto', lg: '75px', xl: ' 75px' } }}
-              align="center"
-              component="th"
-              scope="col"
-            >
-              Count
-            </TableCell>
-          </TableRow>
-        ) : (
-          <TableRow sx={{ height: 0, visibility: 'collapse' }} component="tr">
-            <TableCell component="th" scope="col" aria-label="Empty header" />
-            <TableCell
-              sx={{ width: { md: 'auto', lg: '75px', xl: '75px' } }}
-              component="th"
-              scope="col"
-            />
-          </TableRow>
-        )}
-        {(placement === 'bottomLeft' || placement === 'bottomRight') && (
-          <TableRow
-            sx={{
-              '& th': {
-                border: 'none',
-                backgroundColor: 'transparent',
-                px: 0,
-                pb: 0,
-                fontSize: '14px',
-                ...(placement === 'bottomRight'
-                  ? { color: 'neutrals.white' }
-                  : {})
-              }
-            }}
-            component="tr"
-          >
-            <TableCell align="left" colSpan={2} component="th" scope="col">
-              Open Services with Detected NMI
-            </TableCell>
-          </TableRow>
-        )}
-      </TableHead>
-      <TableBody component="tbody">
-        {data.map((row, idx) => (
-          <TableRow key={idx} sx={tableRowStyles} component="tr">
-            <TableCell sx={{ border: '1px solid #ccc' }} component="td">
-              {row.serviceName}
-            </TableCell>
-            <TableCell
-              sx={{
-                border: '1px solid #ccc',
-                backgroundColor: '#ECF7FF !important'
-              }}
-              align="center"
-              component="td"
-            >
-              {formatDisplayValue(row.count)}
-            </TableCell>
-          </TableRow>
-        ))}
-      </TableBody>
-    </Table>
-  );
-}
-
-export default function RiskyOpenServices({
-  data,
-  detectedServices = 0,
-  isKeyMetricsNull = false
-}: RiskyOpenServicesProps) {
-  const { user } = useAuthContext();
-
-  if (!data || data.length === 0 || detectedServices === 0) {
-    return (
-      <CustomAlert
-        headerMsg="No Matching Results Found"
-        bodyMsg={
-          <>
-            There are no detected services for this organization.
-            {user?.user_type !== 'standard' &&
-              ' Adjust side-panel filters to view results.'}
-          </>
-        }
-        isAlertActive={!isKeyMetricsNull}
-      />
-    );
-  }
-  return (
-    <Box mt={'-10px'}>
-      <Box sx={{ display: { xs: 'none', md: 'block' } }}>
-        <Grid container spacing={3}>
-          <Grid size={{ md: 6 }}>
-            <RiskyServicesTable data={data.slice(0, 4)} />
-          </Grid>
-          <Grid size={{ md: 6 }}>
-            <RiskyServicesTable data={data.slice(4, 7)} />
-          </Grid>
-        </Grid>
-        <Divider sx={{ mt: '10px', mb: '-10px' }} />
-        <Grid container spacing={3}>
-          <Grid size={{ md: 6 }}>
-            <RiskyServicesTable
-              data={data.slice(7, 9)}
-              placement={'bottomLeft'}
-            />
-          </Grid>
-          <Grid size={{ md: 6 }}>
-            <RiskyServicesTable
-              data={data.slice(9, 10)}
-              placement={'bottomRight'}
-            />
-          </Grid>
-        </Grid>
-      </Box>
-      <Box sx={{ display: { xs: 'block', md: 'none' } }}>
-        <RiskyServicesTable data={data.slice(0, 7)} />
-        <Divider sx={{ mt: '10px', mb: '-10px' }} />
-        <RiskyServicesTable data={data.slice(7, 10)} placement={'bottomLeft'} />
-      </Box>
-    </Box>
-  );
-}
+// Re-export from refactored components
+export { default } from './RiskyOpenServices/index';
