+++ frontend/src/pages/VulnerabilityScanDash/RiskyOpenServices/RiskyOpenServicesMain.tsx
@@ -1,0 +1,87 @@
+/**
+ * Main RiskyOpenServices component for displaying vulnerable service data
+ *
+ * Renders risky open services in a responsive grid layout with multiple table sections.
+ * Features include:
+ * - Responsive desktop and mobile layouts
+ * - Multi-section table organization (top/bottom, left/right)
+ *
+ */
+
+// React
+import React from 'react';
+
+// Material-UI Components
+import Box from '@mui/material/Box';
+import Divider from '@mui/material/Divider';
+import Grid from '@mui/material/Grid';
+
+// Hooks
+import { useAuthContext } from 'context';
+
+// Components
+import CustomAlert from '@/components/Dashboard/CustomAlert';
+import { RiskyServicesTable } from './RiskyServicesTable';
+
+// Types
+import { RiskyOpenServicesProps } from './types';
+
+export default function RiskyOpenServices({
+  data,
+  detectedServices = 0,
+  isKeyMetricsNull = false
+}: RiskyOpenServicesProps) {
+  const { user } = useAuthContext();
+
+  if (!data || data.length === 0 || detectedServices === 0) {
+    return (
+      <CustomAlert
+        headerMsg="No Matching Results Found"
+        bodyMsg={
+          <>
+            There are no detected services for this organization.
+            {user?.user_type !== 'standard' &&
+              ' Adjust side-panel filters to view results.'}
+          </>
+        }
+        isAlertActive={!isKeyMetricsNull}
+      />
+    );
+  }
+
+  return (
+    <Box mt={'-10px'}>
+      <Box sx={{ display: { xs: 'none', md: 'block' } }}>
+        <Grid container spacing={3}>
+          <Grid size={{ md: 6 }}>
+            <RiskyServicesTable data={data.slice(0, 4)} />
+          </Grid>
+          <Grid size={{ md: 6 }}>
+            <RiskyServicesTable data={data.slice(4, 7)} />
+          </Grid>
+        </Grid>
+        <Divider sx={{ mt: '10px', mb: '-10px' }} />
+        <Grid container spacing={3}>
+          <Grid size={{ md: 6 }}>
+            <RiskyServicesTable
+              data={data.slice(7, 9)}
+              placement={'bottomLeft'}
+            />
+          </Grid>
+          <Grid size={{ md: 6 }}>
+            <RiskyServicesTable
+              data={data.slice(9, 10)}
+              placement={'bottomRight'}
+            />
+          </Grid>
+        </Grid>
+      </Box>
+
+      <Box sx={{ display: { xs: 'block', md: 'none' } }}>
+        <RiskyServicesTable data={data.slice(0, 7)} />
+        <Divider sx={{ mt: '10px', mb: '-10px' }} />
+        <RiskyServicesTable data={data.slice(7, 10)} placement={'bottomLeft'} />
+      </Box>
+    </Box>
+  );
+}
