+++ frontend/src/pages/VulnerabilityScanDash/RiskyOpenServices/RiskyServicesTable.tsx
@@ -1,0 +1,101 @@
+/**
+ * Individual table component for displaying risky services data
+ *
+ * Renders a single table component as part of the refactored RiskyOpenServices module.
+ *
+ */
+
+// React
+import React from 'react';
+
+// Material-UI Components
+import Table from '@mui/material/Table';
+import TableBody from '@mui/material/TableBody';
+import TableCell from '@mui/material/TableCell';
+import TableHead from '@mui/material/TableHead';
+import TableRow from '@mui/material/TableRow';
+
+// Utils
+import { formatDisplayValue } from '@/utils/stringUtils';
+
+// Local Components
+import { RiskyServicesTableProps } from './types';
+import {
+  tableRowStyles,
+  tableHeaderStyles,
+  tableStyles,
+  bottomRowStyles
+} from './styles';
+
+export function RiskyServicesTable({
+  data,
+  placement = 'top'
+}: RiskyServicesTableProps) {
+  return (
+    <Table sx={tableStyles} component="table">
+      <TableHead component="thead">
+        {placement === 'top' ? (
+          <TableRow sx={tableHeaderStyles} component="tr">
+            <TableCell sx={{ minWidth: '147px' }} component="th" scope="col">
+              Service Name
+            </TableCell>
+            <TableCell
+              sx={{ width: { md: 'auto', lg: '75px', xl: ' 75px' } }}
+              align="center"
+              component="th"
+              scope="col"
+            >
+              Count
+            </TableCell>
+          </TableRow>
+        ) : (
+          <TableRow sx={{ height: 0, visibility: 'collapse' }} component="tr">
+            <TableCell component="th" scope="col" aria-label="Empty header" />
+            <TableCell
+              sx={{ width: { md: 'auto', lg: '75px', xl: '75px' } }}
+              component="th"
+              scope="col"
+            />
+          </TableRow>
+        )}
+        {(placement === 'bottomLeft' || placement === 'bottomRight') && (
+          <TableRow
+            sx={{
+              ...bottomRowStyles,
+              '& th': {
+                ...bottomRowStyles['& th'],
+                ...(placement === 'bottomRight'
+                  ? { color: 'neutrals.white' }
+                  : {})
+              }
+            }}
+            component="tr"
+          >
+            <TableCell align="left" colSpan={2} component="th" scope="col">
+              Open Services with Detected NMI
+            </TableCell>
+          </TableRow>
+        )}
+      </TableHead>
+      <TableBody component="tbody">
+        {data.map((row, idx) => (
+          <TableRow key={idx} sx={tableRowStyles} component="tr">
+            <TableCell sx={{ border: '1px solid #ccc' }} component="td">
+              {row.serviceName}
+            </TableCell>
+            <TableCell
+              sx={{
+                border: '1px solid #ccc',
+                backgroundColor: '#ECF7FF !important'
+              }}
+              align="center"
+              component="td"
+            >
+              {formatDisplayValue(row.count)}
+            </TableCell>
+          </TableRow>
+        ))}
+      </TableBody>
+    </Table>
+  );
+}
