+++ frontend/src/pages/VulnerabilityScanDash/RiskyOpenServices/index.ts
@@ -1,0 +1,19 @@
+/**
+ * Export barrel for RiskyOpenServices component module
+ *
+ * Provides clean, centralized exports for:
+ * - Main and sub-components
+ * - TypeScript interfaces and types
+ * - Maintains backward compatibility with existing imports
+ *
+ */
+
+// Main Component Exports
+export { default } from './RiskyOpenServicesMain';
+export { default as RiskyOpenServices } from './RiskyOpenServicesMain';
+
+// Sub-Component Exports
+export { RiskyServicesTable } from './RiskyServicesTable';
+
+// Type Exports
+export type { RiskyOpenServicesProps, RiskyServicesTableProps } from './types';
