+++ frontend/src/pages/VulnerabilityScanDash/RiskyOpenServices/styles.ts
@@ -1,0 +1,54 @@
+/**
+ * Centralized styling objects for RiskyOpenServices components
+ *
+ * Contains reusable Material-UI sx prop objects for consistent styling.
+ *
+ */
+
+export const tableRowStyles = {
+  borderRadius: 5,
+  borderColor: 'gray',
+  '& td': {
+    borderLeft: '1px solid #ccc',
+    backgroundColor: '#fff',
+    py: '5px',
+    px: 2
+  },
+  '& td:first-of-type': {
+    borderTopLeftRadius: 8,
+    borderBottomLeftRadius: 8
+  },
+  '& td:last-of-type': {
+    borderTopRightRadius: 8,
+    borderBottomRightRadius: 8
+  }
+};
+
+export const tableHeaderStyles = {
+  '& th': {
+    border: 'none',
+    backgroundColor: 'transparent',
+    pb: 0,
+    fontSize: '12px'
+  }
+};
+
+export const tableStyles = {
+  borderCollapse: 'separate',
+  borderSpacing: '0 10px',
+  'thead tr:first-of-type th': {
+    pt: 0
+  },
+  tableLayout: 'fixed',
+  width: '100%'
+};
+
+export const bottomRowStyles = {
+  '& th': {
+    border: 'none',
+    backgroundColor: 'transparent',
+    px: 0,
+    pb: 0,
+    fontSize: '14px'
+  }
+};
