+++ frontend/src/pages/VulnerabilityScanDash/RiskyOpenServices/types.ts
@@ -1,0 +1,24 @@
+/**
+ * TypeScript interfaces and types for RiskyOpenServices components
+ *
+ * Defines the contract and data structures for:
+ * - Main RiskyOpenServices component props
+ * - RiskyServicesTable component props
+ * - Service data handling and configuration options
+ *
+ */
+
+// External Types
+import { ServiceData } from 'types/vuln-scan-stats';
+
+// Component Props Interfaces
+export interface RiskyOpenServicesProps {
+  data: ServiceData[];
+  detectedServices?: string | number;
+  isKeyMetricsNull?: boolean;
+}
+
+export interface RiskyServicesTableProps {
+  data: ServiceData[];
+  placement?: string;
+}
